% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_entropy.R
\name{approx_entropy}
\alias{approx_entropy}
\title{Approximate Entropy}
\usage{
approx_entropy(
  breaks = seq(from = .Machine$double.eps, to = 1, length.out = 4),
  browse = FALSE,
  fig = FALSE
)
}
\arguments{
\item{breaks}{numerical}

\item{browse}{logical, stop within function for inspection}

\item{fig}{plot diagnostic}
}
\value{
a list with fields fun, breaks, and slopes
}
\description{
Internal function for LP algorithm, use mep_solve instead.
}
