% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Balance a dataset}
\usage{
balance(svy, pop, structvars)
}
\arguments{
\item{svy}{a dataset to calculate weights for}

\item{pop}{a dataset representing the whole population}

\item{structvars}{character column(s) to balance for}
}
\value{
a numeric vector of length nrow(svy)
}
\description{
See vignette("balance") for more details.
}
