% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Index a variable}
\usage{
index(x, ...)
}
\arguments{
\item{x}{an object of class "Variab"}

\item{...}{elements of a set, must be convertible to character}
}
\value{
a 0/1 vector of length length(x)
}
\description{
Index a variable
}
