% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_dispose.R
\name{mep_dispose}
\alias{mep_dispose}
\title{Dispose MaxEntProblem}
\usage{
mep_dispose(mep)
}
\arguments{
\item{mep}{an object of class "MaxEntProblem"}
}
\description{
Some backends may create a file. Call this function to disconnect from that file.
}
