% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_getvars.R
\name{mep_getvars}
\alias{mep_getvars}
\title{Extract Solution from MaxEntProblem}
\usage{
mep_getvars(mep, simplify = TRUE, verbose = TRUE)
}
\arguments{
\item{mep}{an object of class "MaxEntProblem"}

\item{simplify}{logical, simplify if possible}

\item{verbose}{logical, diagnostic messages}
}
\value{
when varinfo is set, a data.frame, otherwise a numerical vector
}
\description{
If mep_solve was called successfully, use this function to extract the 
variables from the solution
}
