% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_make.R
\name{mep_make}
\alias{mep_make}
\title{Creates an S3 object for specifications of a Maximum Entropy Problem}
\usage{
mep_make(nvar, ncons, varinfo = NULL, control = list())
}
\arguments{
\item{nvar}{the number of variables}

\item{ncons}{the number of constraints}

\item{varinfo}{optional, a character vector of length(nvar) specifing the variable names}

\item{control}{optional, named list of control variables. See details.}
}
\value{
an S3 object of class MaxEntProblem
}
\description{
You need to specify the number of variables and the number of constraints beforehand. 
The constraints itself are set using mep_set_constraint.
Optionally, you can provide variable names. The control argument allows to specify the storage mode
(storage, either "dense", "sparse" or "sqlite"), the solving algorithm (method, either "BB" or "LP"), 
the maximum number of iterations (maxit, default 25) and the tolerance (tol, default .Machine$double.eps^0.25) and,
in case of storage="sqlite", the name of the database file (dbname, default tempfile()).
}
