% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_set_constraint.R
\name{mep_set_constraint}
\alias{mep_set_constraint}
\title{Set Constraint in MaxEntProblem}
\usage{
mep_set_constraint(mep, j, xt, rhs, indices = NULL)
}
\arguments{
\item{mep}{an object of class "MaxEntProblem"}

\item{j}{integer, index of the constraint}

\item{xt}{integer, value to set}

\item{rhs}{numerical, right-hand side of the constraint}

\item{indices}{integer, index/indices of the variables}
}
\value{
an object of class "MaxEntProblem"
}
\description{
If mep_solve was called successfully, use this function to extract the 
variables from the solution
}
