% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_solve.R
\name{mep_solve}
\alias{mep_solve}
\title{Solve a MaxEnt problem}
\usage{
mep_solve(mep, verbose = TRUE)
}
\arguments{
\item{mep}{an S3 object of class MaxEntProblem. See mep_make}

\item{verbose}{if TRUE, extra messages are printed}
}
\value{
an S3 object of class MaxEntProblem
}
\description{
Calculates the maximum entropy distribution given the constraints.
If the solver succeedes, the returned object has status==0.
}
