% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_solve_BB_sqlite.R
\name{mep_solve_BB_sqlite}
\alias{mep_solve_BB_sqlite}
\title{Solve Maximum entropy Problem for SQLite using BB}
\usage{
mep_solve_BB_sqlite(mep, verbose = TRUE)
}
\arguments{
\item{mep}{an object of class "MaxEntProblem"}

\item{verbose}{logical, diagnostic messages}
}
\value{
an object of class "MaxEntProblem"
}
\description{
Internal function, use mep_solve instead.
}
