% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_solve_LP.R
\name{mep_solve_LP}
\alias{mep_solve_LP}
\title{Solve Maximum entropy Problem for dense/sparse Matrix using Linear Programming}
\usage{
mep_solve_LP(mep, verbose = TRUE)
}
\arguments{
\item{mep}{an object of class "MaxEntProblem"}

\item{verbose}{logical, diagnostic messages}
}
\value{
an object of class "MaxEntProblem"
}
\description{
Internal function, use mep_solve instead.
}
