% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_solve_one.R
\name{mep_solve_one}
\alias{mep_solve_one}
\title{Solve Maximum entropy Problem with one constraint}
\usage{
mep_solve_one(mep, i, only_lambda = TRUE)
}
\arguments{
\item{mep}{an object of class "MaxEntProblem"}

\item{i}{integer, index of the constraint}

\item{only_lambda}{logical, return lambda only (default TRUE)}
}
\value{
an object of class "MaxEntProblem" (only_lambda==FALSE) or numeric
}
\description{
Internal function, use mep_solve instead.
}
