% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mep_test.R
\name{mep_test}
\alias{mep_test}
\title{Runs tests on data integrity}
\usage{
mep_test(fn = NULL, which = c("summary", "fails", "all"), verbose = FALSE)
}
\arguments{
\item{fn}{file name(s) of the tests to run. Default (NULL) runs all tests.}

\item{which}{character, either "summary" (default, aggregates the results),
"fails" (returns tests that failed), or "all"}

\item{verbose}{FALSE (default) silent, TRUE diagnostic messages}
}
\value{
a data.frame
}
\description{
Runs tests on data integrity
}
