% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Slice a vector into a list where each element has at most n elements}
\usage{
slice(x, n)
}
\arguments{
\item{x}{a vector}

\item{n}{integer, maximum length}
}
\value{
a list
}
\description{
Slice a vector into a list where each element has at most n elements
}
