% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variab.R
\name{variab}
\alias{variab}
\title{Define a variable}
\usage{
variab(...)
}
\arguments{
\item{...}{elements of a set, must be convertible to character}
}
\value{
an object of class "Variab"
}
\description{
Define a variable
}
