% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2df.R
\name{vec2df}
\alias{vec2df}
\title{Index a variable}
\usage{
vec2df(varinfo, vec, simplify = TRUE)
}
\arguments{
\item{varinfo}{an object of class "Variab"}

\item{vec}{a numeric vector}

\item{simplify}{logical, simplify result if possible}
}
\value{
a data.frame
}
\description{
Index a variable
}
